v8.2.3-b.1

+++ lib/post.php
@@ -2156,9 +2156,8 @@ if ( ! class_exists( 'WpssoPost' ) ) {
 
 				if ( empty( $shortlink ) ) {
 
-					if ( 'page' === get_post_type( $post_id ) &&
-						$post_id === (int) get_option( 'page_on_front' ) &&
-							'page' === get_option( 'show_on_front' ) ) {
+					if ( 'page' === get_post_type( $post_id ) && (int) $post_id === (int) get_option( 'page_on_front' ) &&
+						'page' === get_option( 'show_on_front' ) ) {
 
 						$shortlink = home_url( '/' );
